教练形象优化

+++ app/src/main/java/com/yijian/staff/bean/CardInfo.kt
@@ -5,10 +5,10 @@ package com.yijian.staff.bean
  * email：850716183@qq.com
  * time: 2018/3/17 19:29:17
  */
-data class CardInfo(val cardprodbaseId: String,val cardType: Int,  val amount: Int, val validDay: Int,   val rechargeGivePercent: Int,val cardName: String,val venusNames: String, val rechargeGiveStr: String, val validTime: Int,val salePrice: Int,    val venueNameList: List<String>) : Comparable<CardInfo> {
+data class CardInfo(val cardprodbaseId: String,val cardType: Integer,  val amount: Integer, val validDay: Integer,   val rechargeGivePercent: Integer,val cardName: String,val venusNames: String, val rechargeGiveStr: String, val validTime: Integer,val salePrice: Integer,    val venueNameList: List<String>) : Comparable<CardInfo> {
 
 
     override fun compareTo(o: CardInfo): Int {
-        return salePrice - o.salePrice
+        return salePrice.toInt() - o.salePrice.toInt()
     }
 }
